requirements.txt
opencv-python
numpy
pillow
scipy